MMOApp/requirements.txt
asgiref==3.8.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==44.0.2
Django==5.1.6
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-filter==25.1
django-js-asset==3.1.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.9
firebase-admin==6.8.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0